Use the first non-vector variable to determine dimensions
I used to use the first variable in the R model.  Which worked well for
6 years!  But the first variable could be a poly, in which case we
iterate through the loop too many times.  Adriana ran into this and it
was causing core dumps.

This change finds the first "non-vector" as I called them 6 years ago
and uses that to determine the size.